requirements.txt
python-dotenv
requests
pandas
google-cloud-bigquery
pytest
apache-airflow-providers-google


